+++ .readthedocs.yaml
@@ -1,0 +1,9 @@
+version: 2
+
+build:
+  os: "ubuntu-22.04"
+  tools:
+    python: "miniconda3-4.7"
+
+conda:
+  environment: environment.yml
