+++ IngeoML/.github/workflows/pip.yaml
@@ -1,0 +1,47 @@
+name: Pip
+
+on: 
+  push:
+    tags:
+      - v*   
+
+jobs:
+  build:
+    runs-on: ${{ matrix.os }}
+    defaults:
+      run:
+        shell: bash -l {0}
+    strategy:
+      matrix:
+        os: [ubuntu-latest, macos-latest, windows-latest]
+        python-version: ["3.8", "3.9", "3.10", "3.11"]
+    steps:
+    - uses: actions/checkout@v2
+    - name: Set up Python
+      uses: conda-incubator/setup-miniconda@v2
+      with:
+        activate-environment: test
+        auto-update-conda: true
+        python-version: ${{ matrix.python-version }}
+        channels: conda-forge
+        allow-softlinks: true
+        channel-priority: flexible
+        show-channel-urls: true	
+    - name: Install dependencies
+      run: |
+        conda install --yes pip
+        pip install twine build
+        conda install --yes numpy scipy scikit-learn nose
+        python -m build
+    - name: Pip
+      if: ${{ runner.os == 'Linux' }}
+      env:
+        TWINE: ${{ secrets.TWINE }}
+      run: |
+        twine upload --skip-existing -u mgraffg -p $TWINE dist/*.tar.gz;
+    - name: Wheel
+      if: ${{ runner.os != 'Linux' }}
+      env:
+        TWINE: ${{ secrets.TWINE }}              
+      run: |
+        twine upload --skip-existing -u mgraffg -p $TWINE dist/*;
