+++ IngeoML/.github/workflows/test.yaml
@@ -1,0 +1,53 @@
+name: Tests
+
+on: 
+  push:
+    branches:
+      - develop
+
+jobs:
+  build:
+    runs-on: ${{ matrix.os }}
+    defaults:
+      run:
+        shell: bash -l {0}      
+    strategy:
+      matrix:
+        os: [ubuntu-latest, macos-latest, windows-latest]
+        python-version: ["3.8", "3.9", "3.10", "3.11"]
+    steps:
+    - uses: actions/checkout@v2
+    - name: Set up Python
+      uses: conda-incubator/setup-miniconda@v2
+      with:
+        activate-environment: test
+        auto-update-conda: true
+        python-version: ${{ matrix.python-version }}
+        channels: conda-forge
+        allow-softlinks: true
+        channel-priority: flexible
+        show-channel-urls: true	
+    - name: Install dependencies
+      run: |
+        conda install --yes pip
+        pip install coverage
+        pip install coveralls
+        conda install --yes numpy scipy scikit-learn nose
+        python setup.py build_ext --inplace
+    - name: Tests on Linux
+      if: ${{ runner.os == 'Linux' }}
+      run: |
+        which python
+        python --version
+        which coverage
+        nosetests --verbose --with-coverage --cover-package=IngeoDash IngeoDash/tests
+    - name: Tests on macOS and Windows
+      if: ${{ runner.os != 'Linux' }}
+      run: |
+        nosetests --verbose IngeoDash/tests
+    - name: coveralls
+      if: ${{ runner.os == 'Linux' }}
+      env:
+        GITHUB_TOKEN: ${{ secrets.GITHUB_TOKEN }}
+      run: |
+        coveralls --service=github
