+++ IngeoML/docs/source/bootstrap.rst
@@ -1,0 +1,38 @@
+.. _bootstrap:
+
+====================================
+Bootstrap
+====================================
+
+
+Citing
+----------
+
+If you find this model useful for any academic/scientific purpose, we would appreciate citations to the following reference:
+  
+.. code:: bibtex
+
+      @InProceedings{10.1007/978-3-031-33783-3_9,
+      author="Nava-Mu{\~{n}}oz, Sergio
+              and Graff Guerrero, Mario
+              and Escalante, Hugo Jair",
+      editor="Rodr{\'i}guez-Gonz{\'a}lez, Ansel Yoan
+              and P{\'e}rez-Espinosa, Humberto
+              and Mart{\'i}nez-Trinidad, Jos{\'e} Francisco
+              and Carrasco-Ochoa, Jes{\'u}s Ariel
+              and Olvera-L{\'o}pez, Jos{\'e} Arturo",
+      title="Comparison of Classifiers in Challenge Scheme",
+      booktitle="Pattern Recognition",
+      year="2023",
+      publisher="Springer Nature Switzerland",
+      address="Cham",
+      pages="89--98",
+      isbn="978-3-031-33783-3"
+      }
+
+
+:py:class:`IngeoML.bootstrap` API
+-------------------------------------
+
+.. autoclass:: IngeoML.bootstrap.StatisticSamples
+   :members:
