--- IngeoML/docs/source/stats_api.rst
@@ -1,11 +1,0 @@
-.. _stats_api:
-
-================================
-:py:class:`IngeoML.stats` API
-================================
-
-.. autoclass:: IngeoML.stats.BootstrapSample
-   :members:
-
-.. autoclass:: IngeoML.stats.CI
-   :members:
