+++ environment.yml
@@ -1,0 +1,5 @@
+name: IngeoML
+channels:
+  - conda-forge
+dependencies:
+  - python
